requirements.txt
beautifulsoup4==4.12.2
torch==2.1.0
transformers==4.35.0
playwright==1.39.0
pyyaml==6.0.1
numpy==1.24.3
pandas==2.1.1
scikit-learn==1.3.1
pytest==7.4.2
pytest-asyncio==0.21.1
loguru==0.7.2
tqdm==4.66.1